Created file twitch_requests with API requests for getting broadcast status
